[Res][P2] Implement Response Parser Component response_parser.py created from part of the api_requester.py dictionaries key names response_parser.py were changed
